Crate bounded_integer

This crate provides two types of bounded integer.

Macro-generated bounded integers

The bounded_integer! macro allows you to define your own bounded integer type, given a specific range it inhabits. For example:

bounded_integer! {
    struct MyInteger { 0..8 }
}
let num = MyInteger::new(5).unwrap();
assert_eq!(num, 5);

This macro supports both structs and enums. See the examples module for the documentation of generated types.

Const generics-based bounded integers

You can also create ad-hoc bounded integers via types in this library that use const generics, for example:

let num = <BoundedU8<0, 7>>::new(5).unwrap();
assert_eq!(num, 5);

These integers are shorter to use as they don’t require a type declaration or explicit name, and they interoperate better with other integers that have different ranges. However due to the limits of const generics, they do not implement some traits like Default.

no_std

All the integers in this crate depend only on libcore and so work in #![no_std] environments.

Crate Features

By default, no crate features are enabled.

std: Interopate with std — implies alloc. Enables the following things:

An implementation of Error for ParseError.

Support for indexing with the const-generic integers on VecDeque.

alloc: Interopate with alloc. Enables the following things:

Support for indexing with the const-generic integers on Vec.

macro: Enable the bounded_integer! macro.

types: Enable the bounded integer types that use const generics.

arbitrary1: Implement Arbitrary for the bounded integers. This is useful when using bounded integers as fuzzing inputs.

bytemuck1: Implement Contiguous for all bounded integers, and Zeroable for macro-generated bounded integers that support it.

num-traits02: Implement Bounded, AsPrimitive, FromPrimitive, NumCast, ToPrimitive, CheckedAdd, CheckedDiv, CheckedMul, CheckedNeg, CheckedRem, CheckedSub, MulAdd, SaturatingAdd, SaturatingMul and SaturatingSub for all const-generic bounded integers.

serde1: Implement Serialize and Deserialize for the bounded integers, making sure all values will never be out of bounds. This has a deprecated alias serde.

zerocopy: Implement IntoBytes for all bounded integers, and Unaligned for suitable macro-generated ones.

step_trait: Implement the Step trait which allows the bounded integers to be easily used in ranges. This will require you to use nightly and place #![feature(step_trait)] in your crate root if you use the macro.

Structs

BoundedI8

An i8 constrained to be in the range MIN..=MAX.

BoundedI16

An i16 constrained to be in the range MIN..=MAX.

BoundedI32

An i32 constrained to be in the range MIN..=MAX.

BoundedI64

An i64 constrained to be in the range MIN..=MAX.

BoundedI128

An i128 constrained to be in the range MIN..=MAX.

BoundedIsize

An isize constrained to be in the range MIN..=MAX.

BoundedU8

An u8 constrained to be in the range MIN..=MAX.

BoundedU16

An u16 constrained to be in the range MIN..=MAX.

BoundedU32

An u32 constrained to be in the range MIN..=MAX.

BoundedU64

An u64 constrained to be in the range MIN..=MAX.

BoundedU128

An u128 constrained to be in the range MIN..=MAX.

BoundedUsize

An usize constrained to be in the range MIN..=MAX.

ParseError

An error which can be returned when parsing a bounded integer.

Enums

ParseErrorKind

The cause of the failure to parse the integer.